Patch Merge Support in prime-mvc in support of RFC 7396. Related to GH FusionAuth Issue #441. https://github.com/FusionAuth/fusionauth-issues/issues/441
This is just a test to prove the algorithm. In theory then all we need to do is support this algorithm in the JSONResult workflow when we see a Content-Type header value of application/merge-patch+json. We could add a new content handler, or re-use this one and bolt on merge support.