Don't require InputStream to be non-null when we are in a HEAD request.
This allows a user to optionally leave this null and not open a file descriptor when in a HEAD request.

Also, close out the InputStream after we transfer it to the OutputStream. I think this was making the assumption that the caller is using a ByteArrayInputStream (for example) that does not need to be closed.