add an action level option to ignore unknown parameters.
In practice, there are use cases where we have a parameter name provided by configuration. This means we cannot know it until runtime. So we end up using the @UnknownParameters map so that we can develop with allowUnknownParameters = false.

Fix some other tests that were using .withParameter in a get() method.

Pull in formatting from fusionauth.